pyproject.toml
[tool.poetry]
name = "tg-assistant-for-cards-dealer"
version = "0.1.0"
description = ""
authors = ["Fobos <95380054+Yastrah@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
